Introduce ResolvedConfig
