+++ crates/server-config/src/lib.rs
@@ -35,21 +35,22 @@ mod kademlia_config;
 mod keys;
 mod listen_config;
 mod network_config;
+mod resolved_config;
 mod services_config;
 
 pub use defaults::{air_interpreter_path, default_base_dir};
-pub use fluence_config::{deserialize_config, load_config};
+pub use resolved_config::{deserialize_config, load_config};
 
 pub use app_services::AppServicesConfig;
 pub use bootstrap_config::BootstrapConfig;
-pub use fluence_config::FluenceConfig;
 pub use kademlia_config::KademliaConfig;
 pub use listen_config::ListenConfig;
 pub use network_config::NetworkConfig;
+pub use resolved_config::ResolvedConfig;
 pub use services_config::ServicesConfig;
 
 pub mod config_keys {
-    pub use crate::fluence_config::{
+    pub use crate::resolved_config::{
         BLUEPRINT_DIR, BOOTSTRAP_NODE, CERTIFICATE_DIR, CONFIG_FILE, EXTERNAL_ADDR, LOCAL,
         MANAGEMENT_PEER_ID, ROOT_KEY_PAIR, ROOT_KEY_PAIR_FORMAT, ROOT_KEY_PAIR_GENERATE,
         ROOT_KEY_PAIR_PATH, SERVICES_WORKDIR, SERVICE_ENVS, TCP_PORT, WEBSOCKET_PORT,
